refactor(test): move TP+MTP test to nightly suite
Move DeepSeek V3.2 TP+MTP test to nightly as a third variant in
test_deepseek_v32_tp.py. Update summary table to include MTP metrics.
